rename screen names and change file structure
