+++ App.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { NavigationContainer } from '@react-navigation/native';
-import Stack from './screens-nav/navigation/Stacks';
+import Stack from './src/navigation/Stacks';
 // import Tabs from './screens-nav/navigation/Tabs';
 
 // screens
-import LoginScreen from './screens-nav/screens/login/LoginScreen';
+import LoginScreen from './src/screens/login/LoginScreen';
 
 export default function App() {
   return (
