+++ src/components/styles.js
