+++ src/navigation/StackNavigator.js
