+++ src/navigation/Tabs.js
